Edit marketplaces and item templates sections
