+++ src/specs/ItemTemplateSpecs.js
@@ -28,24 +28,3 @@ export const ItemTemplateRedeemableOfferSpec = {
     hide_if_expired: false
   }
 };
-
-export const ItemTemplateMediaSectionSpec = {
-  id: undefined,
-  name: "<New Media Section>",
-  collections: []
-};
-
-export const ItemTemplateMediaCollectionSpec = {
-  id: undefined,
-  name: "<New Media Collection>",
-  display: "Media",
-  show_autoplay: false,
-  media: []
-};
-
-export const ItemTemplateReferenceImageSpec = {
-  id: undefined,
-  image_id: "<image_id>",
-  alt_text: "",
-  image: undefined
-};
