+++ src/specs/MarketplaceSpecs.js
@@ -66,192 +66,3 @@ export const MarketplaceCollectionSpec = {
   reveal_animation_mobile: undefined,
   hide_text: false
 };
-
-export const MarketplaceVotingEventSpec = {
-  id: undefined,
-  title: "<New Voting Event>",
-  description: "",
-  exclusive: false,
-  start_date: undefined,
-  end_date: undefined,
-  type: "specified",
-  items: []
-};
-
-export const MarketplaceStorefrontSectionSpec = {
-  id: undefined,
-  name: "<New Storefront Section>",
-  section_header: "",
-  section_subheader: "",
-  section_header_rich_text: "",
-  section_footer: "",
-  type: "Standard",
-  items: [],
-  featured_view_justification: "Left",
-  show_carousel_gallery: false,
-  show_countdown: false,
-  countdown: {
-    header: "",
-    date: undefined
-  }
-};
-
-export const MarketplaceStorefrontBannerSpec = {
-  sku: undefined,
-  image: undefined,
-  image_mobile: undefined,
-  video: undefined,
-  video_muted: true,
-  modal_video: undefined,
-  link: ""
-};
-
-export const MarketplaceFooterLinkSpec = {
-  id: undefined,
-  text: "",
-  url: "",
-  image: "",
-  image_alt: "",
-  content_rich_text: "",
-  content_html: undefined
-};
-
-export const MarketplaceLoginConsentOption = {
-  key: "",
-  message: "",
-  initially_checked: false,
-  required: false
-};
-
-export const MarketplaceSpec = {
-  tenant_id: undefined,
-  tenant_slug: undefined,
-  preview_password_digest: "",
-  branding: {
-    name: "",
-    show: false,
-    external_link: "",
-    notification: {
-      active: false,
-      header: "",
-      text: "",
-    },
-    description: "",
-    header_logo: undefined,
-    header_image: undefined,
-    card_banner_front: undefined,
-    card_banner_back: undefined,
-    background: undefined,
-    background_mobile: undefined,
-    disable_secondary_market: false,
-    hide_leaderboard: false,
-    hide_global_navigation: false,
-    hide_name: false,
-    hide_secondary_in_store: false,
-    disable_usdc: false,
-    use_tenant_styling: false,
-    tabs: {
-      store: "",
-      stores: "",
-      listings: "",
-      my_items: ""
-    },
-    additional_marketplaces: [],
-    tags: [],
-    text_justification: "Left",
-    item_text_justification: "Left",
-    color_scheme: "Light",
-    custom_css: "",
-  },
-  items: [],
-  banners: [],
-  footer_links: [],
-  storefront: {
-    header: "",
-    subheader: "",
-    background: undefined,
-    background_mobile: undefined,
-    show_rich_text_descriptions: false,
-    show_card_cta: false,
-
-    purchase_animation: undefined,
-    purchase_animation_mobile: undefined,
-    reveal_animation: undefined,
-    reveal_animation_mobile: undefined,
-    hide_text: false,
-    skip_reveal: true,
-  },
-  collections_info: {
-    header: "<New Collection>",
-    subheader: "",
-    icon: undefined,
-    banner: undefined,
-    show_on_storefront: false,
-    redeem_animation: undefined,
-    redeem_animation_mobile: undefined,
-    reveal_animation: undefined,
-    reveal_animation_mobile: undefined,
-    hide_text: false
-  },
-  collections: [],
-  voting_events: [],
-  terms: "",
-  terms_document: {
-    link_text: "",
-    terms_document: undefined
-  },
-  login_customization: {
-    logo: undefined,
-    background: undefined,
-    background_mobile: undefined,
-    large_logo_mode: false,
-    log_in_button: {
-      text_color: { color: "" },
-      background_color: { color: "" },
-      border_color: { color: "" }
-    },
-    sign_up_button: {
-      text_color: { color: "" },
-      background_color: { color: "" },
-      border_color: { color: "" }
-    },
-    require_consent: true,
-    default_consent: true,
-    require_email_verification: true,
-    disable_third_party: false,
-    custom_consent: {
-      enabled: false,
-      type: "Modal",
-      consent_modal_header: "",
-      button_text: "",
-      options: []
-    }
-  },
-  payment_options: {
-    wallet_balance: {
-      enabled: true,
-    },
-    stripe: {
-      enabled: true
-    },
-    ebanx: {
-      enabled: false,
-      preferred: false,
-      pix_enabled: false,
-      allowed_countries: []
-    },
-    coinbase: {
-      enabled: true
-    },
-    circle: {
-      enabled: false
-    }
-  },
-  localizations: [],
-  display_currencies: [],
-  default_display_currency: ""
-};
-
-
-
-export default MarketplaceSpec;
